server/requirements.txt
flask
pymysql
flask-cors
bcrypt
pyjwt